Fix capitalization and non-valid move
Made it so that whatever the user entered would be converted into the
correct capitalization so that they had more possible text that they
could enter which would work with the game. Also fixed a problem where
if the user entered nothing, or a string like "blooblah" instead of
a valid move, the game would say that they won. Made it so that it
would return a string telling the user that they didn't enter a valid
move.