Make game() function and change
Made a game function which makes a five round game for the
user to play when it is called. It prompts the user for a move
each round and shows stats for each previous game and also shows
overall stats. At the end of the five rounds it declares if you win,
lose, or tie.

In order to do this the playRound() function had to be reworked
so that instead of returning strings it returned numbers which
would correspond to a win, loss, tie, or non-valid move. This needed
to be done because the game() function wouldn't be able to
collect stats from the old playRound() function. The game() function
could use numbers to collect stats and also send the messages that
the playRound() function was going to. If the user enters a non-valid
move, the function will alert them of this and not count the round
as one of the five.