Increase smartness of javascript and add images
New functions were made in the javascript to decrease the repetition
of statements. An array and forEach method were added to use less code
for adding the eventListeners to the rock, paper, and scissors elements.

Background images were added to the "move selection" elements and the
text was removed. The elements were put in a flexbox display to create
a better looking page. eventListeners were added to detect the motion
of a mouse over the elements, and a function was run in these listeners
which changed the background color.